refactor: migrate pages resolver to standalone field and add page grant criteria utility
